Add support for generating svgs from tex files.
This can be a nice feature when a .tex file contains
a tikz figure.

This is enabled by exposing the dvisvgm binary in the
defined latex toolchain.
And updating the _latex_pdf to support dvi output,
which is provided through a custom Provider,
LatexOutputInfo.

In addition, it also adds a github workflow with a
local bazel cache.