Add support for generating svgs from tex files
This can be a nice feature when a .tex file contains
a tikz figure.

This is enabled by exposing the dvisvgm binary in the
defined latex texlive toolchan.

dvisvgm is dependent on ghostscript as a dynamic lib,
which is why it has been added.

The previously referenced bazel docker image
in cirrus.yaml was not actively updated by google.
The last supported version was 3.5.
So the docker image was changed to the newly
provided image. See:

https://github.com/bazelbuild/continuous-integration/issues/1060
https://github.com/bazelbuild/continuous-integration/pull/1401

ghostscript is built from source using rules_foreign_cc,
which can't be entirely bootstrapped at this point.
(It can't build autoconf)

So it needs that to be installed from e.g. apt.
The bazel docker container did as far as I investigated
not allow to install packages in the container,
I could be wrong though.

Instead I created a github workflow, which depends on bazelisk.
bazelisk is a downloader for bazel. You can configure it always
run with a specific version. But sense cirrus.yaml uses latest,
we don't state any version meaning it will always download the
latest bazel.

It also enables a bazel cache which "should" speed up execution
in the ci pipeline.